refactor(cycle23): split src/logic/tests.rs into 4 domain submodules
- tests.rs: 365 → 5 LOC
- New tests_dir/{users,emails,mailbox,store}.rs, each ≤ 173 LOC
